RF_11.001.02 | MainPageTest - > delete class, rename MainPage2Test to MainPageTest
