+++ src/test/java/school/redrover/GlobalViewTest.java
@@ -22,7 +22,7 @@ public class GlobalViewTest extends BaseTest {
     }
 
     @Test
-    public void testAddDescription() {
+    public void testAddViewDescription() {
         getWait5().until(ExpectedConditions.elementToBeClickable(By.id("description-link"))).click();
         getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.name("description"))).sendKeys(DESC_MESSAGE);
         getWait5().until(ExpectedConditions.elementToBeClickable(By.xpath("//button[@name='Submit']"))).click();
@@ -31,8 +31,8 @@ public class GlobalViewTest extends BaseTest {
                 By.id("description-content"))).getText(), DESC_MESSAGE);
     }
 
-    @Test(dependsOnMethods = "testAddDescription")
-    public void testUpdateDescription() {
+    @Test(dependsOnMethods = "testAddViewDescription")
+    public void testUpdateViewDescription() {
         getWait5().until(ExpectedConditions.elementToBeClickable((By.linkText("Edit description")))).click();
         getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//textarea[@name='description']"))).clear();
         getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//textarea[@name='description']"))).sendKeys(UPDATED_DESC_MESSAGE);
@@ -42,8 +42,8 @@ public class GlobalViewTest extends BaseTest {
         Assert.assertEquals(getDriver().findElement(By.id("description-content")).getText(), UPDATED_DESC_MESSAGE);
     }
 
-    @Test(dependsOnMethods = "testUpdateDescription")
-    public void testCancelUpdateDescription() {
+    @Test(dependsOnMethods = "testUpdateViewDescription")
+    public void testCancelUpdateViewDescription() {
         getWait5().until(ExpectedConditions.elementToBeClickable((By.linkText("Edit description")))).click();
         getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//textarea[@name='description']"))).clear();
         getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//textarea[@name='description']"))).sendKeys(DESC_MESSAGE);
@@ -54,13 +54,24 @@ public class GlobalViewTest extends BaseTest {
                 By.id("description-content"))).getText(), UPDATED_DESC_MESSAGE);
     }
 
-    @Test(dependsOnMethods = "testCancelUpdateDescription")
-    public void testDeleteDescription() {
+    @Test(dependsOnMethods = "testCancelUpdateViewDescription")
+    public void testDeleteViewDescription() {
         getWait5().until(ExpectedConditions.elementToBeClickable(By.id("description-link"))).click();
         getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//textarea[@name='description']"))).clear();
         getWait5().until(ExpectedConditions.elementToBeClickable(By.xpath("//button[@name='Submit']"))).click();
 
         Assert.assertEquals(getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.id("description-link"))).getText(),
                 TEXT_DESCRIPTION_BUTTON);
+    }
+
+    @Test(dependsOnMethods = "testDeleteViewDescription")
+    public void testSaveWithoutViewDescription() {
+        getWait5().until(ExpectedConditions.elementToBeClickable(By.cssSelector("#description-link.jenkins-button"))).click();
+        getWait5().until(ExpectedConditions.elementToBeClickable(By.xpath("//button[@name='Submit']"))).click();
+        getWait10().until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("#description-link.jenkins-button")));
+
+        Assert.assertTrue(
+                getWait5().until(ExpectedConditions.presenceOfElementLocated(By.id("description-content"))).getText().isEmpty(),
+                "Description has non-empty content!");
     }
 }
