+++ src/test/java/school/redrover/MainPageTest.java
@@ -1,66 +1,43 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
+import school.redrover.common.TestUtils;
+
+import java.util.List;
 
 public class MainPageTest extends BaseTest {
 
-    private static final String TEXT_CONTENT = "TEST";
+    private static final String PIPELINE_NAME = "PipelineName";
+    private static final String FOLDER_NAME = "FolderName";
 
-    private void openDescription(){
-        getDriver().findElement(By.cssSelector("#description-link.jenkins-button")).click();
+    private void goToMainPage() {
+        WebElement logo = getWait10().until(ExpectedConditions.elementToBeClickable(By.cssSelector("a.app-jenkins-logo")));
+        ((JavascriptExecutor) getDriver()).executeScript("arguments[0].click();", logo);
+        getWait10().until(ExpectedConditions.visibilityOfElementLocated(
+                By.xpath("//a[@href='/view/all/newJob']"))).isDisplayed();
     }
 
-    @Ignore
     @Test
-    public void testAddDescription() {
-        openDescription();
+    public void testOrderName() {
+        TestUtils.createJob(getDriver(), getWait10(), PIPELINE_NAME, TestUtils.JobType.PIPELINE);
+        goToMainPage();
+        TestUtils.createJob(getDriver(), getWait10(), FOLDER_NAME, TestUtils.JobType.FOLDER);
+        goToMainPage();
 
-        getDriver().findElement(By.name("description")).sendKeys(TEXT_CONTENT);
-        getDriver().findElement(By.name("Submit")).click();
+        getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//span[text()='" + PIPELINE_NAME + "']")));
+        getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//span[text()='" + FOLDER_NAME + "']")));
 
-        getWait5().until(ExpectedConditions.
-                visibilityOfElementLocated(By.cssSelector("#description-link.jenkins-button")));
+        List<String> actualOrder = getDriver().findElements(By.cssSelector(".jenkins-table__link.model-link.inside"))
+                .stream()
+                .map(WebElement::getText)
+                .toList();
 
-        Assert.assertEquals(
-                getDriver().findElement(By.id("description-content")).getText(),
-                TEXT_CONTENT);
-    }
-
-    @Ignore
-    @Test(dependsOnMethods = "testAddDescription")
-    public void testChangeDescription() {
-        String changedDescription = TEXT_CONTENT + "_changed";
-
-        openDescription();
-
-        WebElement descriptionField = getDriver().findElement(By.name("description"));
-        descriptionField.clear();
-        descriptionField.sendKeys(changedDescription);
-        getDriver().findElement(By.name("Submit")).click();
-
-        getWait5().until(ExpectedConditions.
-                visibilityOfElementLocated(By.cssSelector("#description-link.jenkins-button")));
-
-        Assert.assertEquals(
-                getDriver().findElement(By.id("description-content")).getText(),
-                changedDescription);
-    }
-
-    @Ignore
-    @Test
-    public void testSaveWithoutDescription() {
-        openDescription();
-        getDriver().findElement(By.name("Submit")).click();
-
-        getWait10().until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("#description-link.jenkins-button")));
-        Assert.assertTrue(
-                getDriver().findElement(By.id("description-content")).getText().isEmpty(),
-                "Description has non-empty content!");
+        Assert.assertEquals(actualOrder, actualOrder.stream().sorted().toList(), "Not an alphabetical order!");
     }
 }
